+++ controllers/contacts.js
@@ -1,11 +1,5 @@
-const Joi = require("joi");
 const { HttpError, ctrlWrapper } = require("../utils");
 const contacts = require("../models/contacts");
-const addContactSchema = Joi.object({
-    name: Joi.string().required(),
-    email: Joi.string().required(),
-    phone: Joi.string().required(),
-});
 
 const getAllContacts = async (req, res) => {
     const contactsList = await contacts.listContacts();
@@ -22,10 +16,6 @@ const getContactById = async (req, res) => {
 };
 
 const addContact = async (req, res) => {
-    const { error } = addContactSchema.validate(req.body);
-    if (error) {
-        throw HttpError(400, error.message);
-    }
     const contact = await contacts.addContact(req.body);
     res.status(201).json(contact);
 };
@@ -40,10 +30,6 @@ const removeContact = async (req, res) => {
 };
 
 const updateContact = async (req, res) => {
-    const { error } = addContactSchema.validate(req.body);
-    if (error) {
-        throw HttpError(400, error.message);
-    }
     const { contactId } = req.params;
     const contactUpDate = await contacts.updateContact(contactId, req.body);
     if (!contactUpDate) {
